fix: fix unit tests and timestamp issue
