+++ src/converters.ts
@@ -5,6 +5,8 @@ const dateFormatter = new Intl.DateTimeFormat('en-US', {
   year: 'numeric',
   month: 'numeric',
   day: 'numeric',
+  hour: 'numeric',
+  minute: '2-digit',
 });
 
 const formatTimestamp = (timestamp: Timestamp | null | undefined, fallback = 'Unknown') => {
