+++ src/routes/approved.test.tsx
@@ -41,6 +41,19 @@ vi.mock('../components/Table', () => ({
   },
 }));
 
+vi.mock('../hooks/useTableFilters', () => ({
+  useTableFilters: () => ({
+    columnFilters: [],
+    setColumnFilters: vi.fn(),
+    renderTextFilterControl: () => null,
+    renderSelectFilterControl: () => null,
+    renderDateRangeFilterControl: () => null,
+  }),
+  caseInsensitiveIncludesFilter: vi.fn(),
+  dateRangeFilter: vi.fn(),
+  mrrcFilter: vi.fn(),
+}));
+
 describe('approved.tsx', () => {
   beforeEach(() => {
     mocks.navigateMock.mockReset();
